+++ cmd/template/cluster/provider/azure.go
@@ -36,10 +36,6 @@ func WriteAzureTemplate(out io.Writer, config ClusterCRsConfig) error {
 		infrastructureRef := newCAPZClusterInfraRef(azureClusterCR)
 
 		clusterCR := newCAPIV1Alpha3ClusterCR(config, infrastructureRef)
-		{
-			// XXX: azure-operator reconciles Cluster & MachinePool to set OwnerReferences (for now).
-			clusterCR.GetLabels()[label.AzureOperatorVersion] = config.ReleaseComponents["azure-operator"]
-		}
 		clusterCRYaml, err = yaml.Marshal(clusterCR)
 		if err != nil {
 			return microerror.Mask(err)
@@ -81,7 +77,6 @@ func newAzureClusterCR(config ClusterCRsConfig) *capzv1alpha3.AzureCluster {
 			Name:      config.ClusterID,
 			Namespace: config.Namespace,
 			Labels: map[string]string{
-				label.AzureOperatorVersion:    config.ReleaseComponents["azure-operator"],
 				label.Cluster:                 config.ClusterID,
 				capiv1alpha3.ClusterLabelName: config.ClusterID,
 				label.Organization:            config.Owner,
@@ -111,7 +106,6 @@ func newAzureMasterMachineCR(config ClusterCRsConfig) *capzv1alpha3.AzureMachine
 			Name:      fmt.Sprintf("%s-master-%d", config.ClusterID, 0),
 			Namespace: config.Namespace,
 			Labels: map[string]string{
-				label.AzureOperatorVersion:                config.ReleaseComponents["azure-operator"],
 				label.Cluster:                             config.ClusterID,
 				capiv1alpha3.ClusterLabelName:             config.ClusterID,
 				capiv1alpha3.MachineControlPlaneLabelName: "true",